Actions not saved yet.

+++ screens/HomeScreen.js
@@ -1107,6 +1107,11 @@ class Home extends Component {
 												})
 											}}
 											onPress={async () => {
+												if (await AsyncStorage.getItem("isFetchingNotCompleated") == "true") {
+													// Actions not saved yet
+													return;
+												}
+
 												await this.databaseRepository.clear();
 												await AsyncStorage.clear();
